requirements.txt
# Requirements for Microsoft Agent Framework Lab

# Core packages
agent-framework
azure-identity>=1.15.0
azure-ai-projects>=1.0.0
azure-ai-agents>=1.0.0
azure-core>=1.28.0

# Azure Service Bus SDK
azure-servicebus>=7.12.0

# Azure Cosmos DB SDK (async)
aiohttp>=3.9.0
azure-cosmos>=4.7.0

# Azure Document Intelligence (Form Recognizer)
azure-ai-documentintelligence>=1.0.0

# FastAPI Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
jinja2>=3.1.0

# Microsoft Graph SDK for email access
msgraph-sdk>=1.0.0

# Optional but recommended
python-dotenv>=1.0.0  # For .env file support
